backend/requirements.txt
fastapi
uvicorn
faster-whisper
python-multipart
python-dotenv
aiofiles
medcat
spacy
twilio